use zerolog

+++ internal/install/install.go
@@ -4,35 +4,63 @@ import (
 	"bytes"
 	"context"
 	"fmt"
-	"io"
 	"os"
 	"text/template"
 
+	"github.com/golgoth31/release-installer/internal/output"
 	"github.com/golgoth31/release-installer/internal/release"
 	getter "github.com/hashicorp/go-getter"
-	log "github.com/sirupsen/logrus"
+
+	logger "github.com/golgoth31/release-installer/internal/log"
 	"github.com/spf13/viper"
-	"github.com/vbauerster/mpb/v5"
-	"github.com/vbauerster/mpb/v5/decor"
 )
 
 var yamlData *viper.Viper
 var releaseData *release.Release
 var defaultProgressBar getter.ProgressTracker = &progressBar{}
 
-func NewInstall(rel string) *Install {
-	yamlData = viper.New()
-	releaseData = release.New(rel)
-	return &Install{ApiVersion: "release/v1", Kind: "Install"}
-}
+var out output.Output
 
 func (i *Install) LoadYaml(file string) {
 	yamlData.SetConfigType("yaml")
 	yamlData.SetConfigFile(file)
 
 	if err := yamlData.ReadInConfig(); err != nil {
-		log.Fatalf("Failed to read %s", file)
+		// out.Status(out.FatalStatus(), fmt.Sprintf("Failed to read %s", file))
+		logger.StdLog.Fatal().Err(err).Msg("")
 	}
+}
+func (i *Install) templates() (
+	releaseURL bytes.Buffer,
+	releaseFileName bytes.Buffer,
+	checksumURL bytes.Buffer,
+	checksumFileName bytes.Buffer) {
+	// template strings
+	treleaseURL := template.Must(template.New("releaseURL").Parse(releaseData.Spec.Url))
+	if err := treleaseURL.Execute(&releaseURL, i.Spec); err != nil {
+		// out.Status(out.FatalStatus(), "Error templating release URL")
+		logger.StdLog.Fatal().Err(err).Msg("")
+	}
+
+	treleaseFileName := template.Must(template.New("releaseFileName").Parse(releaseData.Spec.File.Archive))
+	if err := treleaseFileName.Execute(&releaseFileName, i.Spec); err != nil {
+		// out.Status(out.FatalStatus(), "Error templating release file name")
+		logger.StdLog.Fatal().Err(err).Msg("")
+	}
+
+	tchecksumURL := template.Must(template.New("checksumURL").Parse(releaseData.Spec.Checksum.Url))
+	if err := tchecksumURL.Execute(&checksumURL, i.Spec); err != nil {
+		// out.Status(out.FatalStatus(), "Error templating checksum URL")
+		logger.StdLog.Fatal().Err(err).Msg("")
+	}
+
+	tchecksumFileName := template.Must(template.New("checksumFileName").Parse(releaseData.Spec.Checksum.File))
+	if err := tchecksumFileName.Execute(&checksumFileName, i.Spec); err != nil {
+		// out.Status(out.FatalStatus(), "Error templating checksum file name")
+		logger.StdLog.Fatal().Err(err).Msg("")
+	}
+
+	return releaseURL, releaseFileName, checksumURL, checksumFileName
 }
 
 func (i *Install) Install() {
@@ -42,52 +70,35 @@ func (i *Install) Install() {
 	link := i.Spec.Path + "/" + releaseData.Spec.File.Binary
 	file := link + "_" + i.Spec.Version
 
-	// template strings
-	treleaseURL := template.Must(template.New("releaseURL").Parse(releaseData.Spec.Url))
-	var breleaseURL bytes.Buffer
-	if err := treleaseURL.Execute(&breleaseURL, i.Spec); err != nil {
-		log.Fatal("Error templating release URL")
-	}
-	treleaseFileName := template.Must(template.New("releaseFileName").Parse(releaseData.Spec.File.Archive))
-	var breleaseFileName bytes.Buffer
-	if err := treleaseFileName.Execute(&breleaseFileName, i.Spec); err != nil {
-		log.Fatal("Error templating release file name")
-	}
-	tchecksumURL := template.Must(template.New("checksumURL").Parse(releaseData.Spec.Checksum.Url))
-	var bchecksumURL bytes.Buffer
-	if err := tchecksumURL.Execute(&bchecksumURL, i.Spec); err != nil {
-		log.Fatal("Error templating checksum URL")
-	}
-	tchecksumFileName := template.Must(template.New("checksumFileName").Parse(releaseData.Spec.Checksum.File))
-	var bchecksumFileName bytes.Buffer
-	if err := tchecksumFileName.Execute(&bchecksumFileName, i.Spec); err != nil {
-		log.Fatal("Error templating checksum file name")
-	}
+	releaseURL, releaseFileName, checksumURL, checksumFileName := i.templates()
 
 	downURL := fmt.Sprintf(
 		"%s/%s",
-		breleaseURL.String(),
-		breleaseFileName.String(),
+		releaseURL.String(),
+		releaseFileName.String(),
 	)
 	getterDownURL := fmt.Sprintf(
 		"%s?checksum=file:%s/%s",
 		downURL,
-		bchecksumURL.String(),
-		bchecksumFileName.String(),
+		checksumURL.String(),
+		checksumFileName.String(),
 	)
 
-	log.Infof("Downloading archive file")
+	out.StepTitle("Release files")
 	fmt.Println()
-	log.Infof("Archive file: %s", downURL)
-	log.Infof("Checksum file: %s", fmt.Sprintf(
-		"%s/%s",
-		bchecksumURL.String(),
-		bchecksumFileName.String(),
-	))
+	logger.StdLog.Info().Msgf("Checksum file: %s/%s",
+		checksumURL.String(),
+		checksumFileName.String(),
+	)
+	logger.StdLog.Info().Msgf("Archive file:  %s", downURL)
+
+	fmt.Println()
+	out.StepTitle("Downloading files")
+	fmt.Println()
 
 	pwd, err := os.Getwd()
 	if err != nil {
-		log.Fatalf("Error getting wd: %v", err)
+		logger.StdLog.Fatal().Msgf("Error getting wd: %v", err)
 	}
 
 	ctx, cancel := context.WithCancel(context.Background())
@@ -104,66 +115,28 @@ func (i *Install) Install() {
 	}
 
 	if err := client.Get(); err != nil {
-		log.Fatal(err)
+		logger.StdLog.Fatal().Err(err).Msg("")
 	}
 
 	// ensure the binary is executable
 	if err := os.Chmod(file, 0755); err != nil {
-		log.Fatal(err)
+		logger.StdLog.Fatal().Err(err).Msg("")
 	}
+
 	fmt.Println()
-	log.Info("Done")
-	log.Infof("File saved as: %s", file)
-	fmt.Println()
+	logger.StdLog.Info().Msgf("File saved as: %s", file)
 
 	if i.Spec.Default {
-		log.Infof("Creating symlink: %s", link)
-		if err := os.Remove(link); err != nil {
-			log.Fatal(err)
+		logger.StdLog.Info().Msgf("Creating symlink: %s", link)
+		_, err := os.Stat(link)
+		if err == nil {
+			if err := os.Remove(link); err != nil {
+				logger.StdLog.Fatal().Err(err).Msg("")
+			}
 		}
 		if err := os.Symlink(file, link); err != nil {
-			log.Fatal(err)
+			logger.StdLog.Fatal().Err(err).Msg("")
 		}
-		log.Info("Done")
+		logger.StdLog.Info().Msgf("Done")
 	}
 }
-
-func (cpb *progressBar) TrackProgress(src string, currentSize, totalSize int64, stream io.ReadCloser) io.ReadCloser {
-	cpb.lock.Lock()
-	defer cpb.lock.Unlock()
-
-	pb := mpb.New(mpb.WithWidth(60))
-	// Parameters of th new progress bar
-	bar := pb.AddBar(totalSize,
-		mpb.PrependDecorators(
-			decor.Name(src),
-			decor.Name(" "),
-			decor.CountersKiloByte("% .2f / % .2f "),
-			decor.AverageSpeed(decor.UnitKB, "(% .2f)"),
-		),
-		mpb.AppendDecorators(
-			decor.Percentage(),
-			decor.Name(" - "),
-			decor.Elapsed(decor.ET_STYLE_GO, decor.WC{W: 4}),
-			decor.Name(" - "),
-			decor.OnComplete(
-				decor.AverageETA(decor.ET_STYLE_GO, decor.WC{W: 4}), "done",
-			),
-		),
-	)
-
-	reader := bar.ProxyReader(stream)
-
-	return &readCloser{
-		Reader: reader,
-		close: func() error {
-			cpb.lock.Lock()
-			defer cpb.lock.Unlock()
-
-			pb.Wait()
-			return nil
-		},
-	}
-}
-
-func (c *readCloser) Close() error { return c.close() }
